requirements.txt
streamlit>=1.34.0
python-dotenv>=1.0.1
pypdf>=4.2.0
faiss-cpu>=1.8.0
tiktoken>=0.7.0

langchain>=0.2.12
langchain-community>=0.2.11
langchain-openai>=0.1.20
langchain-text-splitters>=0.2.2
langchain-core>=0.2.30
